requirements.txt
streamlit==1.41.0
google-generativeai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==11.2.1
urllib3<2.0.0
watchdog
